reorganized the codebase a fair bit
