+++ Common/ModPlayers/NinjaEvadeChancePlayer.cs
