+++ Common/UI/CustomInventory.cs
@@ -25,99 +25,6 @@ using TerrariaCells.Common.Items;
 
 namespace TerrariaCells.Common.UI;
 
-[Autoload(Side = ModSide.Client)]
-public class DeadCellsUISystem : ModSystem
-{
-    static readonly string[] filtered_layers =
-    [
-        "Vanilla: Laser Ruler",
-        "Vanilla: Ruler",
-        "Vanilla: Inventory",
-        "Vanilla: Info Accessories Bar",
-        "Vanilla: Hotbar",
-    ];
-
-    internal UserInterface userInterface;
-    internal LimitedStorageUI limitedStorageUI;
-
-    public override void Load()
-    {
-        if (Main.dedServ)
-        {
-            return;
-        }
-
-        userInterface = new UserInterface();
-        limitedStorageUI = new LimitedStorageUI();
-        limitedStorageUI.Activate();
-        userInterface.SetState(limitedStorageUI);
-    }
-
-    public override void Unload()
-    {
-        limitedStorageUI = null;
-    }
-
-    public override void UpdateUI(GameTime gameTime)
-    {
-        if (!DevConfig.Instance.EnableInventoryChanges)
-        {
-            return;
-        }
-
-        if (userInterface?.CurrentState == null)
-        {
-            userInterface.Update(gameTime);
-        }
-    }
-
-    public void HideVanillaInventoryLayers(List<GameInterfaceLayer> layers)
-    {
-        if (!DevConfig.Instance.HideVanillaInventory)
-        {
-            return;
-        }
-
-        layers.RemoveAll(
-            delegate(GameInterfaceLayer layer)
-            {
-                return filtered_layers.Contains(layer.Name);
-            }
-        );
-    }
-
-    public override void ModifyInterfaceLayers(List<GameInterfaceLayer> layers)
-    {
-        // called here before the filtering of vanilla inventory layers since somethign dumb happens when filtering and idk what
-        ModContent.GetInstance<UISystem>().ModifyInterfaceLayers(layers);
-
-        HideVanillaInventoryLayers(layers);
-
-        int mouseTextIndex = layers.FindIndex(layer => layer.Name.Equals("Vanilla: Mouse Text"));
-        if (mouseTextIndex != -1)
-        {
-            if (!DevConfig.Instance.EnableInventoryChanges)
-            {
-                return;
-            }
-
-            layers.Insert(
-                mouseTextIndex,
-                new LegacyGameInterfaceLayer(
-                    "TerraCells: InventoryUI",
-                    delegate
-                    {
-                        LimitedStorageUI.CustomGUIHotbarDrawInner(); // draws hotbar, aka inventory closed
-                        LimitedStorageUI.CustomDrawInterface_27_Inventory(); // draws inventory, aka inventory open
-                        return true;
-                    },
-                    InterfaceScaleType.UI
-                )
-            );
-        }
-    }
-}
-
 public class CustomAccessorySlot : ModAccessorySlot
 {
     public override bool DrawDyeSlot => false;
